Moving testing from assert calls to internal test function. Fixed return type of tile function from double to int

+++ lab_code/apex_code.h
@@ -6,7 +6,7 @@
 #ifndef APEXLAB_APEX_CODE_H
 #define APEXLAB_APEX_CODE_H
 
-double getNumberTiles(double heightInFt, double widthInFt, int tileSizeInInches);
+int getNumberTiles(double heightInFt, double widthInFt, int tileSizeInInches);
 bool upperCharAnd(unsigned char char1, unsigned char char2);
 
 #endif //APEXLAB_APEX_CODE_H
